% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docker.R
\name{build_image}
\alias{build_image}
\title{Run \code{docker build} in a directory}
\usage{
build_image(machine_name, dir = getwd(), params = "")
}
\arguments{
\item{machine_name}{character. Existing Docker Machine machine_name.}

\item{dir}{character. Directory that contains a Dockerfile}

\item{params}{character. Additional parameters for \code{docker build}}
}
\description{
Run \code{docker build} in a directory
}

