% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.R
\name{machine_AWS}
\alias{machine_AWS}
\title{Create a docker machine on AWS.}
\usage{
machine_AWS(name, vpc_id, cores = 2, ami = "ami-4ae27e22", root_size = 16,
  security_group = "docker-machine", zone = "a", iam = NULL,
  region = NULL, session_token = NULL, subnet_id = NULL)
}
\arguments{
\item{name}{character. Required. The name of the machine.}

\item{vpc_id}{character. Required. The VPC id. Get it from AWS console.}

\item{cores}{integer. Number of cores. More cores = more expensive!}

\item{ami}{character. AMI to use. Don't touch if you don't know what you are doing.
If you do, then see https://docs.docker.com/machine/#drivers for more details.}

\item{root_size}{integer. The root disk size of the instance (in GB). Default: 16}

\item{security_group}{character. AWS VPC security group name. Default: docker-machine}

\item{zone}{character.  The AWS zone launch the instance in (i.e. one of a,b,c,d,e). Default: a}

\item{iam}{character. IAM role for the instance.}

\item{region}{character. EC2 region. Default is us-east-1.}

\item{session_token}{character. Your session token for the Amazon Web Services API.}

\item{subnet_id}{character. AWS VPC subnet id.}
}
\description{
Defaults are pretty sensible. You must provide only the following:
\itemize{
  \item name. Name of the docker machine.
  \item cores. Number of cores. Defaults to 2.
  \item vpc_id. The id of your VPC.
  \item set options(dokk.AWS_ACCESS_KEY_ID) to your AWS access key.
  \item set options(dokk.AWS_SECRET_KEY) to your AWS secret.
}
}
\details{
Note that this can be expensive, so make sure you clean up after yourself!

Creates an EC2 instance for you.
}

