% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/machine.R
\name{machine_config}
\alias{machine_config}
\title{Config.}
\usage{
machine_config(name)
}
\arguments{
\item{name}{character. Name of the machine.}
}
\description{
Show the Docker client configuration for a machine.
}

