% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.R
\name{machine_DigitalOcean}
\alias{machine_DigitalOcean}
\title{Create a docker machine on Digital Ocean.}
\usage{
machine_DigitalOcean(name, size = "2gb", image = NULL, region = NULL,
  ipv6 = NULL, private_networking = NULL, backups = NULL)
}
\arguments{
\item{name}{character. Required. The name of the machine.}

\item{size}{character. The size of the droplet. More memory - more expensive!}

\item{image}{character. The name of the Digital Ocean image to use. Default: docker}

\item{region}{character. The region to create the droplet in, see \code{Regions API} for how to get a list. Default: nyc3}

\item{ipv6}{character. Enable IPv6 support for the droplet. Default: false}

\item{private_networking}{logical. Enable private networking support for the droplet. Default: false}

\item{backups}{logical. Enable Digital Oceans backups for the droplet. Default: false}
}
\description{
Defaults are pretty sensible. You must provide only the following:
\itemize{
  \item name. Name of the docker machine.
  \item size. The size of the droplet. Defaults to 2gb.
  \item set options(dokk.DIGITAL_OCEAN_ACCESS_TOKEN) to your digitalocean access token.
}
}
\details{
Note that this can be expensive, so make sure you clean up after yourself!

Creates a digital ocean droplet for you.
The DigitalOcean driver will use ubuntu-14-04-x64 as the default image.
}

