% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/machine.R
\name{machine_restart}
\alias{machine_restart}
\title{restart.}
\usage{
machine_restart(name)
}
\arguments{
\item{name}{character. Name of the machine.}
}
\description{
Restart a machine. Oftentimes this is equivalent
to \code{docker-machine stop; machine start}.
}

