% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/machine.R
\name{machine_rm}
\alias{machine_rm}
\title{Remove docker machine.}
\usage{
machine_rm(name)
}
\arguments{
\item{name}{character. Name of the machine.}
}
\description{
Remove a machine. This will remove the
local reference as well as delete it on the cloud provider or
virtualization management platform.
}

