% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docker.R
\name{run_image}
\alias{run_image}
\title{Run your container in the docker machine}
\usage{
run_image(machine_name, image, params = "", command = "")
}
\arguments{
\item{machine_name}{character. Existing Docker Machine machine_name.}

\item{image}{character. A hash or a tag of docker image to be run.}

\item{params}{character. Additional parameters for \code{docker run}, i.e. \code{-ti}.}

\item{command}{character. The command to be executed in the container. Leave blank for default CMD.}
}
\description{
Run your container in the docker machine
}

