% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docker.R
\name{build_image}
\alias{build_image}
\title{Run \code{docker build} in a directory}
\usage{
build_image(machine_name, dir = ".", params = "", intern = TRUE)
}
\arguments{
\item{machine_name}{character. Existing Docker Machine machine_name.}

\item{dir}{character. Directory that contains a Dockerfile}

\item{params}{character. Additional parameters for \code{docker build}}

\item{intern}{logical. If set to TRUE the building process will be invisible,
and the function will return the hash of the built image. If false,
the build steps will be shown, but the function will return the exit status}
}
\description{
Run \code{docker build} in a directory
}

