% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/machine.R
\name{machine_local}
\alias{machine_local}
\title{Start a local docker machine, driver = virtualbox}
\usage{
machine_local(name)
}
\arguments{
\item{name}{character. Name of the machine.}
}
\description{
Start a local docker machine, driver = virtualbox
}

